# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/createLabel": {
            "post": {
                "summary": "create label",
                "description": "create user label",
                "operationId": "5fa93c835d46a9760f3ccb556db3e8ac",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "labelname"
                                ],
                                "properties": {
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label Added Sucessfully"
                    },
                    "409": {
                        "description": "Label Name Already Exists"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "202": {
                        "description": "Label Not Added"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/readAllLabels": {
            "get": {
                "summary": "Read All Labels for an User",
                "description": "Read User Labels",
                "operationId": "28df23d3d8a0c2554673aa3ff50154d4",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "Labels Retrieved Successfully."
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    },
                    "404": {
                        "description": "Notes Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateLabel": {
            "post": {
                "summary": "update label",
                "description": "update user label",
                "operationId": "9bd4f40296813d35395547ede2055562",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "labelname"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label Updated Successfully"
                    },
                    "202": {
                        "description": "Label Not Updated"
                    },
                    "404": {
                        "description": "Label Not Found"
                    },
                    "409": {
                        "description": "Label Name Already Exists"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteLabel": {
            "post": {
                "summary": "Delete Label",
                "description": "Delete User Label",
                "operationId": "9c3cba02e360bf9c17688fd5b53ccc67",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label Successfully Deleted"
                    },
                    "404": {
                        "description": "Label Not Found"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}